SLACK_EXPOSE_CHANNEL_NAME
Pull request #2 isn't correct.

This patch renames SLACK_EXPOSE_CHANNEL_NAME_AS_FROM env var to
SLACK_EXPOSE_CHANNEL_NAME. Also, previous patch (#2) touches to
`message.from`, but this patch makes it to touch `message.to`.

Message's `from` and `to` is supposed to be like an email header.

Let's see other existing adapter.

- XMPP based adapter (e.g. ruboty-slack):

  In channels, `from` is a combination of room ID and user ID,
  and `to` is a room ID.

  In private messages, `from` and `to` are both user ID.

And our previous behavior is this:

  In channels, `from` is a room ID. `from_name` is a user name.
  And `to` is a room ID.

Okay. Ideally `from` should be a combination of room and user ID... But
it breaks existing stuff. So this patch reverts `from` behavior to
before #2, and makes `to` customizable.